pyproject.toml
[tool.poetry]
name = "repl_python3_Troll-Bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.7.3"
replit = "^3.2.4"
Flask = "^2.0.1"
wolframalpha = "^5.0.0"
youtube_dl = "^2021.6.6"
tweepy = "^4.1.0"
HTMLParser = "^0.0.2"
PyNaCl = "^1.4.0"
agent-admin = "^0.1.1"
reactions = "^0.0.1"
xkcd = "^2.4.2"
pyfiglet = "^0.8.post1"
utils = "^1.0.1"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
